More riscv-test. Bug fixes.
    - Added other tests.
    - Updated how the instruction rom is loaded in verilog.
    - Added code to also load the data ram (required for load tests)
    - Fixed bugs for jalr instruction.

+++ tests/cocotb/riscv-test-simple/run_all.py
@@ -24,7 +24,43 @@ import shutil
 
 # isa and its run time
 rv32ui_isa = [
+    ['add'   , 10000],
+    ['addi'  , 10000 ],
+    ['and'   , 10000],
+    ['andi'  , 10000 ],
+    ['auipc'  , 10000 ],
+    ['beq'  , 10000 ],
+    ['bge'  , 10000 ],
+    ['bgeu'  , 10000 ],
+    ['blt'  , 10000 ],
+    ['bltu'  , 10000 ],
+    ['bne'  , 10000 ],
+    #['fence_i'  , 10000 ],
+    ['jal'  , 10000 ],
+    ['jalr'  , 10000 ],     # ??? Having x
+    ['lb'   , 10000],       # ??? Having x
+    ['lbu'  , 10000 ],      # ??? Having x
+    ['lh'   , 10000],       # ??? Having x
+    ['lhu'  , 10000 ],      # ??? Having x
+    ['lui'  , 10000 ],
+    ['lw'  , 10000 ],
+    ['or'   , 10000],
+    ['ori'  , 10000 ],
+    ['sb'   , 10000],
+    ['sh'  , 10000 ],
     ['simple', 2000 ],
+    ['sll'   , 10000],
+    ['slli'  , 10000 ],
+    ['slt'   , 10000],
+    ['slti'  , 10000 ],
+    ['sltu'   , 10000],
+    ['sltiu'  , 10000 ],
+    ['sra'   , 10000],
+    ['srai'  , 10000 ],
+    ['srl'   , 10000],
+    ['srli'  , 10000 ],
+    ['sub'   , 10000],
+    ['sw'  , 10000 ],
     ['xor'   , 10000],
     ['xori'  , 3000 ],
 ]
@@ -53,7 +89,9 @@ RESULT = 'results.xml'
 
 def clear_all():
     """ Clear all the output """
-    shutil.rmtree(OUTPUT_DIR)
+    if os.path.isdir(OUTPUT_DIR):
+        shutil.rmtree(OUTPUT_DIR)
+    os.system("make clean_all")
 
 def move_result(test, arch):
     """ Move the test result to its own directory """
@@ -79,6 +117,7 @@ def check_result():
 
 def one_test(test, arch, runtime):
     """ run all the process for one tests """
+    os.system("make clean_rom")
     run_test(test, arch, runtime)
     result = check_result()
     move_result(test, arch)
@@ -106,14 +145,18 @@ def print_result(results):
         False: "FAIL"
     }
     print("=======================================")
-    print("==            Tests Result           ==")
+    print("              Tests Result             ")
     print("=======================================")
     for arch in results.keys():
         print("ISA: " + arch)
         tests = results[arch]
         for test in tests.keys():
             rst = translate[tests[test]]
-            print(test + ": " + str(rst))
+            print(test + ": " + str(rst), end='')
+            if rst == 'PASS':
+                print()
+            else:
+                print('    <---- ' + test + ' ---->')
 
 
 def main():
